refactor(s2n-quic-platform): move socket config to separate module
