+++ quic/s2n-quic-platform/src/lib.rs
@@ -16,4 +16,5 @@ pub mod features;
 pub mod io;
 pub mod message;
 pub mod socket;
+mod syscall;
 pub mod time;
